인식 좀 하고 저장되게 수정
